refactor(script): split pet/homun/merc buildins into buildin_pet.c
Issue #61 sixth step: move pet, homunculus, and mercenary helpers out of script.c.
